[KOGITO-848] Change kogito DMN editor to extend from kie-wb-common-kogito-webapp-base (#3103)
* [KOGITO-848] Extend from kie-wb-common-kogito-webapp-base inside DMN editor

* [KOGITO-848] Add dependency to fix tests

* [KOGITO-848] Fixed dependency order. Fixed runtime gwt:run

* [KOGITO-848] Add missing dependency

* [KOGITO-848] Add missing dependency
