--- kie-wb-common-dmn/kie-wb-common-dmn-webapp-kogito-common/src/main/java/org/kie/workbench/common/dmn/webapp/kogito/common/backend/workarounds/DummyAuthenticationService.java
@@ -1,48 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.workbench.common.dmn.webapp.kogito.common.backend.workarounds;
-
-import javax.enterprise.context.Dependent;
-
-import org.jboss.errai.bus.server.annotations.Service;
-import org.jboss.errai.security.shared.api.identity.User;
-import org.jboss.errai.security.shared.api.identity.UserImpl;
-import org.jboss.errai.security.shared.service.AuthenticationService;
-
-@Service
-@Dependent
-public class DummyAuthenticationService implements AuthenticationService {
-
-    @Override
-    public void logout() {
-    }
-
-    @Override
-    public User login(String username, String password) {
-        return new UserImpl("admin");
-    }
-
-    @Override
-    public boolean isLoggedIn() {
-        return true;
-    }
-
-    @Override
-    public User getUser() {
-        return new UserImpl("admin");
-    }
-}
