+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp-kogito-marshaller-tests/pom.xml
@@ -80,6 +80,10 @@
     </dependency>
     <dependency>
       <groupId>org.kie.workbench</groupId>
+      <artifactId>kie-wb-common-kogito-webapp-base</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.workbench</groupId>
       <artifactId>kie-wb-common-dmn-webapp-kogito-marshaller</artifactId>
     </dependency>
     <dependency>
