+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp-kogito-testing/pom.xml
@@ -181,11 +181,6 @@
       <groupId>org.uberfire</groupId>
       <artifactId>uberfire-preferences-api</artifactId>
     </dependency>
-    <!-- Required to provide default implementation of server-side CDI injection points -->
-    <dependency>
-      <groupId>org.uberfire</groupId>
-      <artifactId>uberfire-preferences-backend</artifactId>
-    </dependency>
     <!-- Required to provide default implementation of client-side CDI injection points -->
     <dependency>
       <groupId>org.uberfire</groupId>
@@ -239,6 +234,15 @@
       <groupId>org.kie.workbench</groupId>
       <artifactId>kie-wb-common-dmn-webapp-kogito-marshaller</artifactId>
       <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.workbench</groupId>
+      <artifactId>kie-wb-common-kogito-webapp-base</artifactId>
+      <classifier>sources</classifier>
+    </dependency>
+    <dependency>
+      <groupId>org.kie.workbench</groupId>
+      <artifactId>kie-wb-common-kogito-webapp-base</artifactId>
     </dependency>
     <dependency>
       <groupId>org.uberfire</groupId>
@@ -421,6 +425,11 @@
       <groupId>org.kie.workbench.stunner</groupId>
       <artifactId>kie-wb-common-stunner-kogito-client</artifactId>
       <scope>provided</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.uberfire</groupId>
+      <artifactId>appformer-kogito-bridge</artifactId>
+      <classifier>sources</classifier>
     </dependency>
 
     <dependency>
