+++ kie-wb-common-dmn/kie-wb-common-dmn-webapp-kogito-testing/src/main/java/org/kie/workbench/common/dmn/showcase/client/perspectives/TestingPerspectiveConfiguration.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2019 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.kie.workbench.common.dmn.showcase.client.perspectives;
+
+import javax.enterprise.inject.Alternative;
+
+import org.kie.workbench.common.kogito.webapp.base.client.perspectives.PerspectiveConfiguration;
+import org.uberfire.client.workbench.panels.impl.AbstractWorkbenchPanelPresenter;
+import org.uberfire.client.workbench.panels.impl.MultiListWorkbenchPanelPresenter;
+
+@Alternative
+public class TestingPerspectiveConfiguration extends PerspectiveConfiguration {
+
+    @Override
+    public Class<? extends AbstractWorkbenchPanelPresenter> getPerspectivePanelType() {
+        return MultiListWorkbenchPanelPresenter.class;
+    }
+}
