+++ kie-wb-common-kogito/kie-wb-common-kogito-webapp-base/src/test/java/org/kie/workbench/common/kogito/webapp/base/client/workarounds/IsKogitoTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,15 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.kie.workbench.common.kogito.webapp.base.client.workarounds;
 
-package org.kie.workbench.common.dmn.webapp.kogito.common.client.workarounds;
+import org.junit.Before;
+import org.junit.Test;
 
-import javax.enterprise.inject.Alternative;
+import static org.junit.Assert.assertTrue;
 
-@Alternative
-public class IsKogito extends org.kie.workbench.common.widgets.client.kogito.IsKogito {
+public class IsKogitoTest {
 
-    public boolean get() {
-        return true;
+    private IsKogito isKogito;
+
+    @Before
+    public void setup() {
+        this.isKogito = new IsKogito();
+    }
+
+    @Test
+    public void testIsKogito() {
+        assertTrue(isKogito.get());
     }
 }
